+++ apibuilder4elastic/custom_flow_nodes/api-builder-plugin-axway-api-management/src/actions.js
@@ -77,7 +77,14 @@ async function lookupCurrentUser(params, options) {
 			throw new Error('The requestHeaders do not contain the required header csrf-token');
 		}
 		logger.trace(`Trying to get current user based on VIDUSR cookie.`);
-		user.loginName = await _getCurrentGWUser(headers = {'Cookie': requestHeaders.cookie});
+		try {
+			user.loginName = await _getCurrentGWUser(headers = {'Cookie': requestHeaders.cookie});
+		} catch (err) { 
+			// Might happen if the request has been sent to the wrong ANM by a Load-Balancer in between. (Session Stickyness not working as expected)
+			// Only mitigating the problem, but not really fully solving the issue - Load-Balanced request must be investigated
+			logger.warn(`Unexpected error while trying to get current user from the ANM. Using a Load-Balancer which is not sticky?! Try again at least once.`);
+			user.loginName = await _getCurrentGWUser(headers = {'Cookie': requestHeaders.cookie});
+		}
 		logger.trace(`Current user is: ${user.loginName}`);
 		permissions = await _getCurrentGWPermissions(headers = {'Cookie': requestHeaders.cookie, 'csrf-token': requestHeaders['csrf-token']}, user.loginName);
 	}